requirements.txt
Django==3.2.9
psycopg2==2.9.3
Pillow==9.0.1
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.21.0